refactor: apply code style, package and others (#55)
